further cleanup on TC doppler proxy
